Fixed compilation warnings/errors

+++ Core/SNES_SPC_misc.cpp
@@ -304,7 +304,7 @@ void SNES_SPC::set_output( sample_t* out, int size )
 			assert( out <= out_end );
 		}
 		
-		dsp.set_output( out, out_end - out );
+		dsp.set_output( out, (int)(out_end - out) );
 	}
 	else
 	{
